+++ e2e-tests/ui-components/tests/cardDetails.spec.js
@@ -7,7 +7,7 @@ test.describe("card component", () => {
   });
 
   Object.values(VALID_CARDS).forEach((card) => {
-    test(`can capture valid ${card.brand} card details`, async ({ page }) => {
+    test(`can capture valid card details for ${card.number} (${card.brand})`, async ({ page }) => {
       let values = {};
 
       await page.exposeFunction("handleChange", (newValues) => {
@@ -26,6 +26,7 @@ test.describe("card component", () => {
       await frame.getByLabel("CVC").fill(card.cvc);
       await expect.poll(async () => values.card?.number).toBeEncrypted();
       await expect.poll(async () => values.card?.brand).toEqual(card.brand);
+      await expect.poll(async () => values.card?.localBrands).toEqual(card.localBrands);
       await expect.poll(async () => values.card?.cvc).toBeEncrypted();
       await expect
         .poll(async () => values.card?.expiry?.month)
@@ -64,6 +65,34 @@ test.describe("card component", () => {
     await frame.getByLabel("Number").blur();
     await expect(frame.getByText("Your card number is invalid")).toBeVisible();
     await expect.poll(async () => values.errors.number).toEqual("invalid");
+    await expect.poll(async () => values.isValid).toBeFalsy();
+    await expect(
+      frame.getByText("Your expiration date is invalid")
+    ).not.toBeVisible();
+    await expect(frame.getByText("Your CVC is invalid")).not.toBeVisible();
+  });
+
+  test("shows an error message for an unsupported card brand", async ({
+    page,
+  }) => {
+    let values = {};
+
+    await page.exposeFunction("handleChange", (newValues) => {
+      values = newValues;
+    });
+
+    await page.evaluate(() => {
+      const card = window.evervault.ui.card({acceptedBrands: ["visa"]});
+      card.on("change", window.handleChange);
+      card.mount("#form");
+    });
+
+    const testCard = VALID_CARDS.amex;
+    const frame = page.frameLocator("iframe[data-evervault]");
+    await frame.getByLabel("Number").fill(testCard.number);
+    await frame.getByLabel("Number").blur();
+    await expect(frame.getByText("This card brand is not supported")).toBeVisible();
+    await expect.poll(async () => values.errors.number).toEqual("unsupportedBrand");
     await expect.poll(async () => values.isValid).toBeFalsy();
     await expect(
       frame.getByText("Your expiration date is invalid")