+++ packages/ui-components/src/Card/CardNumber.tsx
@@ -1,4 +1,4 @@
-import cardValidator from "card-validator";
+import { validateNumber } from "@evervault/card-validator";
 import { FocusEvent, useEffect, useRef } from "react";
 import { useMask } from "../utilities/useMask";
 
@@ -43,7 +43,7 @@ export function CardNumber({
     ] as CardMask[],
     dispatch: (appended, dynamicMasked) => {
       const number = dynamicMasked.value + appended;
-      const brand = cardValidator.number(number).card?.type;
+      const { brand } = validateNumber(number);
 
       const mask = dynamicMasked.compiledMasks.find((m) => {
         const maskBrand = (m as CardMask).brand;