+++ packages/ui-components/src/Card/types.ts
@@ -1,10 +1,11 @@
-import type { ThemeObject, CardField, CardTranslations } from "types";
+import type { ThemeObject, CardField, CardTranslations, CardBrandName } from "types";
 
 export interface CardConfig {
   theme?: ThemeObject;
   autoFocus?: boolean;
   hiddenFields?: ("number" | "expiry" | "cvc")[];
   fields?: CardField[];
+  acceptedBrands?: CardBrandName[];
   translations?: Partial<CardTranslations>;
 }