+++ e2e-apps/decrypt-backend/CHANGELOG.md
@@ -143,3 +143,15 @@
 ### Patch Changes
 
 - @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null