+++ examples/custom-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-custom-theme
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [a96e24b]
+  - @evervault/browser@2.23.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [43d5607]
+  - @evervault/browser@2.22.1
+
 ## 0.0.4
 
 ### Patch Changes