+++ examples/forms/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-forms
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [a96e24b]
+  - @evervault/browser@2.23.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [43d5607]
+  - @evervault/browser@2.22.1
+
 ## 0.0.2
 
 ### Patch Changes